refactor: AuthService 위치 변경
